Fixed poll struct type

+++ src/safeposix/syscalls/net_constants.rs
@@ -336,12 +336,12 @@ pub const MINSOCKOBJID: i32 = 0;
 pub const MAXSOCKOBJID: i32 = 1024;
 
 //POLL CONSTANTS
-pub const POLLIN: u32 = 01;  // There is data to read.
-pub const POLLPRI: u32 = 02; //There is urgent data to read.
-pub const POLLOUT: u32 = 04; // Writing now will not block.
-pub const POLLERR: u32 = 010; // Error condition.
-pub const POLLHUP: u32 = 020; // Hung up.
-pub const POLLNVAL: u32 = 040; // Invalid polling request.
+pub const POLLIN: i16 = 01;  // There is data to read.
+pub const POLLPRI: i16 = 02; //There is urgent data to read.
+pub const POLLOUT: i16 = 04; // Writing now will not block.
+pub const POLLERR: i16 = 010; // Error condition.
+pub const POLLHUP: i16 = 020; // Hung up.
+pub const POLLNVAL: i16 = 040; // Invalid polling request.
 
 //EPOLL CONSTANTS
 pub const EPOLLIN: i32 = 0x001;
